# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserRole": {
				"enum": [
					"ADMIN",
					"USER"
				],
				"type": "string"
			},
			"MediaType": {
				"enum": [
					"USER_PROFILE",
					"USER_CITIZENSHIP",
					"USER_PASSPORT",
					"BLOG_THUMBNAIL"
				],
				"type": "string"
			},
			"Media": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"mimeType": {
						"type": "string"
					},
					"mediaType": {
						"$ref": "#/components/schemas/MediaType"
					}
				},
				"required": [
					"id",
					"createdAt",
					"deletedAt",
					"updatedAt",
					"name",
					"mimeType",
					"mediaType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"isVerified": {
						"type": "boolean"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"profile": {
						"$ref": "#/components/schemas/Media"
					}
				},
				"required": [
					"id",
					"createdAt",
					"deletedAt",
					"updatedAt",
					"fullName",
					"email",
					"password",
					"isVerified",
					"role",
					"profile"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterUserValidation": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					}
				},
				"required": [
					"fullName",
					"email",
					"password",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InitializePasswordResetValidator": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FinalizePasswordResetValidator": {
				"properties": {
					"email": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginValidator": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GenerateAccessToken": {
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-nodejs-starter",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user": {
			"get": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"ADMIN",
							"USER"
						]
					}
				],
				"parameters": []
			}
		},
		"/user/all": {
			"get": {
				"operationId": "GetAllUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalCount": {
											"type": "number",
											"format": "double"
										},
										"lastPage": {
											"type": "number",
											"format": "double"
										},
										"currentPage": {
											"type": "number",
											"format": "double"
										},
										"nextPage": {
											"type": "number",
											"format": "double"
										},
										"prevPage": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/User"
											},
											"type": "array"
										}
									},
									"required": [
										"totalCount",
										"lastPage",
										"currentPage",
										"nextPage",
										"prevPage",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/UserRole"
						}
					}
				]
			}
		},
		"/auth/register-user": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterUserValidation"
							}
						}
					}
				}
			}
		},
		"/auth/email-verification/{userId}/{token}": {
			"get": {
				"operationId": "EmailVerification",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InitializePasswordResetValidator"
							}
						}
					}
				}
			}
		},
		"/auth/finalize-forgot-password/{token}": {
			"post": {
				"operationId": "FinalizePasswordReset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FinalizePasswordResetValidator"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"role": {},
												"refreshToken": {},
												"accessToken": {},
												"id": {},
												"isVerified": {
													"type": "boolean"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"isVerified",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"message": {},
												"role": {
													"$ref": "#/components/schemas/UserRole"
												},
												"refreshToken": {},
												"accessToken": {},
												"isVerified": {
													"type": "boolean",
													"enum": [
														true
													],
													"nullable": false
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"role",
												"refreshToken",
												"accessToken",
												"isVerified",
												"id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginValidator"
							}
						}
					}
				}
			}
		},
		"/auth/generate-access-token": {
			"post": {
				"operationId": "GenerateAccessToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refreshToken": {},
										"accessToken": {}
									},
									"required": [
										"refreshToken",
										"accessToken"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenerateAccessToken"
							}
						}
					}
				}
			}
		},
		"/media": {
			"post": {
				"operationId": "Upload",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Media"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												},
												"status": {
													"type": "string"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									},
									"mediaType": {
										"type": "string"
									}
								},
								"required": [
									"file",
									"mediaType"
								]
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:4000",
			"description": "Development server"
		}
	]
}